Исправление выбора цвета
