+++ LamerHelper.csproj
@@ -51,4 +51,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Modules\Tools\" />
+  </ItemGroup>
+
 </Project>
