--- Modules/Feature/ColorPickerModule/ColorPickerModule.xaml
@@ -1,54 +1,0 @@
-<UserControl x:Class="LamerHelper.Modules.Feature.ColorPickerModule"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
-             xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
-             Height="Auto" Width="Auto">
-    <Grid>
-        <ikw:SimpleStackPanel Spacing="10">
-            <Button x:Name="buttonPickColor" Click="ButtonPickColor_Click">
-                <ui:IconAndText Icon="{x:Static ui:SegoeFluentIcons.Eyedropper}"
-                              Content="Выбрать цвет" />
-            </Button>
-            
-            <Grid Visibility="Collapsed" x:Name="colorInfoPanel">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="Auto"/>
-                    <ColumnDefinition Width="*"/>
-                </Grid.ColumnDefinitions>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="Auto"/>
-                </Grid.RowDefinitions>
-                
-                <!-- Образец цвета -->
-                <Border Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="2"
-                        x:Name="colorSample" 
-                        Height="50" 
-                        BorderBrush="Black" 
-                        BorderThickness="1" 
-                        Margin="0,0,0,10"/>
-                
-                <!-- Значения цвета -->
-                <TextBlock Grid.Row="1" Grid.Column="0" Text="HEX:" Margin="0,0,10,0" VerticalAlignment="Center"/>
-                <TextBox Grid.Row="1" Grid.Column="1" x:Name="textBoxHex" IsReadOnly="True"/>
-                
-                <TextBlock Grid.Row="2" Grid.Column="0" Text="RGB:" Margin="0,0,10,0" VerticalAlignment="Center"/>
-                <TextBox Grid.Row="2" Grid.Column="1" x:Name="textBoxRgb" IsReadOnly="True"/>
-                
-                <!-- Кнопка копирования -->
-                <StackPanel Grid.Row="3" Grid.Column="0" Grid.ColumnSpan="2" Orientation="Horizontal" Margin="0,10,0,0">
-                    <Button x:Name="buttonCopyHex" Content="Копировать HEX" Click="ButtonCopyHex_Click" Margin="0,0,10,0"/>
-                    <Button x:Name="buttonCopyRgb" Content="Копировать RGB" Click="ButtonCopyRgb_Click"/>
-                </StackPanel>
-            </Grid>
-            
-            <TextBlock x:Name="textBlockInstructions" 
-                       Text="Нажмите кнопку 'Выбрать цвет', затем наведите мышь на нужный цвет и нажмите левую кнопку мыши для захвата." 
-                       TextWrapping="Wrap"
-                       Margin="0,5,0,0"/>
-        </ikw:SimpleStackPanel>
-    </Grid>
-</UserControl>
