+++ Modules/Tools/ColorPickerModule/ColorPickerModule.xaml
@@ -1,0 +1,42 @@
+<UserControl x:Class="LamerHelper.Modules.Tools.ColorPickerModule"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
+             xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
+             Height="Auto" Width="Auto">
+    <Grid Margin="10">
+        <ikw:SimpleStackPanel Spacing="10">
+            <Button x:Name="ButtonPickColor" Click="ButtonPickColor_Click">
+                <ui:IconAndText Icon="{x:Static ui:SegoeFluentIcons.Eyedropper}"
+                                Content="Выбрать цвет"/>
+            </Button>
+            <Grid x:Name="ColorInfoPanel" Visibility="Collapsed">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+                <Border x:Name="ColorSample" Height="50"
+                        BorderBrush="Black" BorderThickness="1"
+                        Margin="0,0,0,10" Grid.Row="0"
+                        CornerRadius="6"/>
+                <StackPanel Orientation="Horizontal" Grid.Row="1" Margin="0,0,0,5">
+                    <TextBlock Text="HEX:" FontWeight="Bold" Margin="0,0,10,0" VerticalAlignment="Center"/>
+                    <TextBox x:Name="TextBoxHex" IsReadOnly="True" VerticalAlignment="Center"/>
+                </StackPanel>
+                <StackPanel Orientation="Horizontal" Grid.Row="2" Margin="0,0,0,5">
+                    <TextBlock Text="RGB:" FontWeight="Bold" Margin="0,0,10,0" VerticalAlignment="Center"/>
+                    <TextBox x:Name="TextBoxRgb" IsReadOnly="True" VerticalAlignment="Center"/>
+                </StackPanel>
+                <StackPanel Orientation="Horizontal" Grid.Row="3" Margin="0,10,0,0">
+                    <Button x:Name="ButtonCopyHex" Content="Копировать HEX" Click="ButtonCopyHex_Click" Margin="0,0,10,0"/>
+                    <Button x:Name="ButtonCopyRgb" Content="Копировать RGB" Click="ButtonCopyRgb_Click"/>
+                </StackPanel>
+            </Grid>
+            <TextBlock x:Name="TextBlockInstructions" 
+                       Text="Нажмите кнопку 'Выбрать цвет', затем перемещайте курсор для предпросмотра и нажмите левую кнопку мыши для захвата цвета. Нажмите правую – чтобы отменить."
+                       TextWrapping="Wrap"/>
+        </ikw:SimpleStackPanel>
+    </Grid>
+</UserControl>
